+++ plugin-gradle/src/main/java/com/diffplug/gradle/spotless/KotlinGradleExtension.java
@@ -22,6 +22,7 @@ import java.util.Objects;
 import com.diffplug.common.collect.ImmutableSortedMap;
 import com.diffplug.spotless.FormatterStep;
 import com.diffplug.spotless.kotlin.KtLintStep;
+import com.diffplug.spotless.kotlin.KtfmtStep;
 
 public class KotlinGradleExtension extends FormatExtension {
 	private static final String GRADLE_KOTLIN_DSL_FILE_EXTENSION = "*.gradle.kts";
@@ -62,6 +63,33 @@ public class KotlinGradleExtension extends FormatExtension {
 
 		private FormatterStep createStep() {
 			return KtLintStep.createForScript(version, GradleProvisioner.fromProject(getProject()), userData);
+		}
+	}
+
+	/** Uses the [ktfmt](https://github.com/facebookincubator/ktfmt) jar to format source code. */
+	public KtfmtConfig ktfmt() {
+		return ktfmt(KtfmtStep.defaultVersion());
+	}
+
+	/**
+	 * Uses the given version of [ktfmt](https://github.com/facebookincubator/ktfmt) to format source
+	 * code.
+	 */
+	public KtfmtConfig ktfmt(String version) {
+		Objects.requireNonNull(version);
+		return new KtfmtConfig(version);
+	}
+
+	public class KtfmtConfig {
+		final String version;
+
+		KtfmtConfig(String version) {
+			this.version = Objects.requireNonNull(version);
+			addStep(createStep());
+		}
+
+		private FormatterStep createStep() {
+			return KtfmtStep.create(version, GradleProvisioner.fromProject(getProject()));
 		}
 	}